+++ semcore/d3-chart/CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 
 - Added radius of curvature for `Bar, HorizontalBar`.
 - Added dynamic position tooltip for `HoverRect, HoverLine`.
+- Added hover for `Donut.Pie`.
 
 ## [1.2.0] - 2021-09-24